laptop_ai_guard/requirements.txt
qai-hub-models[cavaface]
opencv-python
numpy
pillow
pyserial
